new class for sending emails
